+++ cookie/src/main/java/com/cookie/domain/review/controller/ReviewController.java
@@ -1,8 +1,8 @@
 package com.cookie.domain.review.controller;
 
-import com.cookie.domain.review.dto.CreateReviewDto;
+import com.cookie.domain.review.dto.request.CreateReviewRequest;
 import com.cookie.domain.review.dto.response.ReviewResponse;
-import com.cookie.domain.review.dto.UpdateReviewDto;
+import com.cookie.domain.review.dto.request.UpdateReviewRequest;
 import com.cookie.domain.review.service.ReviewService;
 import com.cookie.global.util.ApiUtil;
 import com.cookie.global.util.ApiUtil.ApiSuccess;
@@ -20,15 +20,15 @@ public class ReviewController {
     private final ReviewService reviewService;
 
     @PostMapping("/{userId}")
-    public ApiSuccess<?> createReview(@PathVariable(name = "userId") Long userId, @RequestBody CreateReviewDto createReviewDto) {
+    public ApiSuccess<?> createReview(@PathVariable(name = "userId") Long userId, @RequestBody CreateReviewRequest createReviewRequest) {
         // TODO: userId JWT 토큰으로 변경
-        reviewService.createReview(userId, createReviewDto);
+        reviewService.createReview(userId, createReviewRequest);
         return ApiUtil.success("SUCCESS");
     }
 
     @PutMapping("/{reviewId}")
-    public ApiSuccess<?> updateReview(@PathVariable(name = "reviewId") Long reviewId, @RequestBody UpdateReviewDto updateReviewDto) {
-        reviewService.updateReview(reviewId, updateReviewDto);
+    public ApiSuccess<?> updateReview(@PathVariable(name = "reviewId") Long reviewId, @RequestBody UpdateReviewRequest updateReviewRequest) {
+        reviewService.updateReview(reviewId, updateReviewRequest);
         return ApiUtil.success("SUCCESS");
     }
