+++ cookie/src/main/java/com/cookie/domain/review/dto/request/UpdateReviewRequest.java
@@ -1,4 +1,4 @@
-package com.cookie.domain.review.dto;
+package com.cookie.domain.review.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateReviewDto {
+public class UpdateReviewRequest {
     private String content;
     private double movieScore;
     private Boolean isSpoiler;
