+++ cookie/src/main/java/com/cookie/domain/review/service/ReviewService.java
@@ -2,9 +2,9 @@ package com.cookie.domain.review.service;
 
 import com.cookie.domain.movie.entity.Movie;
 import com.cookie.domain.movie.repository.MovieRepository;
-import com.cookie.domain.review.dto.CreateReviewDto;
+import com.cookie.domain.review.dto.request.CreateReviewRequest;
 import com.cookie.domain.review.dto.response.ReviewResponse;
-import com.cookie.domain.review.dto.UpdateReviewDto;
+import com.cookie.domain.review.dto.request.UpdateReviewRequest;
 import com.cookie.domain.review.entity.Review;
 import com.cookie.domain.review.repository.ReviewRepository;
 import com.cookie.domain.user.entity.User;
@@ -25,12 +25,12 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
 
     @Transactional
-    public void createReview(Long userId, CreateReviewDto createReviewDto) {
+    public void createReview(Long userId, CreateReviewRequest createReviewRequest) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("not found userId: " + userId));
         log.info("Retrieved user: userId = {}", userId);
 
-        Long movieId = createReviewDto.getMovieId();
+        Long movieId = createReviewRequest.getMovieId();
         Movie movie = movieRepository.findById(movieId)
                 .orElseThrow(() -> new IllegalArgumentException("not found movieId: " + movieId));
         log.info("Retrieved movie: movieId = {}", movieId);
@@ -39,18 +39,18 @@ public class ReviewService {
             throw new IllegalArgumentException("해당 영화에 이미 리뷰를 등록했습니다.");
         }
 
-        Review review = createReviewDto.toEntity(user, movie);
+        Review review = createReviewRequest.toEntity(user, movie);
         reviewRepository.save(review);
         log.info("Created review: userId = {}, movieId = {}", userId, movieId);
     }
 
     @Transactional
-    public void updateReview(Long reviewId, UpdateReviewDto updateReviewDto) {
+    public void updateReview(Long reviewId, UpdateReviewRequest updateReviewRequest) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new IllegalArgumentException("not found reviewId: " + reviewId));
         log.info("Retrieved review: reviewId = {}", reviewId);
 
-        review.update(updateReviewDto.getContent(), updateReviewDto.getMovieScore(), updateReviewDto.getIsSpoiler());
+        review.update(updateReviewRequest.getContent(), updateReviewRequest.getMovieScore(), updateReviewRequest.getIsSpoiler());
         log.info("Updated review: reviewId = {}", reviewId);
     }
